+++ library/models/sdxl_model_prep.py
@@ -11,7 +11,8 @@ from library.utils.torch_utils import match_mixed_precision
 from library.models import sdxl_original_unet, model_util, sdxl_model_util
 from library.models.model_prep import set_padding_mode_for_vae_conv2d_modules
 from library.config.dataclasses.model import ModelConfig
-from library.config.dataclasses.performance import MemoryConfig, CachingConfig, PrecisionConfig
+from library.config.dataclasses.performance import MemoryConfig, PrecisionConfig
+from library.config.dataclasses.data import CachingConfig
 
 setup_logging()
 logger = logging.getLogger(__name__)